:truck: api: move placement profile service from lib to src
