+++ api/lib/infrastructure/repositories/participant-results-shared-repository.js
@@ -1,7 +1,7 @@
 import { knex } from '../../../db/knex-database-connection.js';
 import { ParticipantResultsShared } from '../../../lib/domain/models/ParticipantResultsShared.js';
+import * as placementProfileService from '../../../src/shared/domain/services/placement-profile-service.js';
 import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
-import * as placementProfileService from '../../domain/services/placement-profile-service.js';
 import * as campaignRepository from './campaign-repository.js';
 
 async function _fetchKnowledgeElements(campaignParticipationId) {
