+++ api/scripts/certification/get-user-certification-eligibility.js
@@ -6,11 +6,11 @@ import * as dotenv from 'dotenv';
 dotenv.config({ path: `${__dirname}/../.env` });
 
 import { disconnect } from '../../db/knex-database-connection.js';
-import * as placementProfileService from '../../lib/domain/services/placement-profile-service.js';
 import { usecases } from '../../lib/domain/usecases/index.js';
 import { learningContentCache as cache } from '../../lib/infrastructure/caches/learning-content-cache.js';
 import { temporaryStorage } from '../../lib/infrastructure/temporary-storage/index.js';
 import * as certificationBadgesService from '../../src/certification/shared/domain/services/certification-badges-service.js';
+import * as placementProfileService from '../../src/shared/domain/services/placement-profile-service.js';
 import { logger } from '../../src/shared/infrastructure/utils/logger.js';
 
 /**
