+++ api/src/prescription/campaign-participation/infrastructure/repositories/campaign-profile-repository.js
@@ -1,6 +1,6 @@
 import { knex } from '../../../../../db/knex-database-connection.js';
 import { NotFoundError } from '../../../../../lib/domain/errors.js';
-import * as placementProfileService from '../../../../../lib/domain/services/placement-profile-service.js';
+import * as placementProfileService from '../../../../shared/domain/services/placement-profile-service.js';
 import * as areaRepository from '../../../../shared/infrastructure/repositories/area-repository.js';
 import * as competenceRepository from '../../../../shared/infrastructure/repositories/competence-repository.js';
 import { CampaignProfile } from '../../domain/models/CampaignProfile.js';
