+++ api/src/shared/domain/services/placement-profile-service.js
@@ -1,19 +1,19 @@
 import bluebird from 'bluebird';
 import _ from 'lodash';
 
+import { PlacementProfile } from '../../../../lib/domain/models/PlacementProfile.js';
+import { UserCompetence } from '../../../../lib/domain/models/UserCompetence.js';
+import * as knowledgeElementRepository from '../../../../lib/infrastructure/repositories/knowledge-element-repository.js';
+import * as knowledgeElementSnapshotRepository from '../../../../lib/infrastructure/repositories/knowledge-element-snapshot-repository.js';
 import {
   CERTIFICATION_VERSIONS,
   CertificationVersion,
-} from '../../../src/certification/shared/domain/models/CertificationVersion.js';
-import * as scoringService from '../../../src/evaluation/domain/services/scoring/scoring-service.js';
-import * as assessmentRepository from '../../../src/shared/infrastructure/repositories/assessment-repository.js';
-import * as assessmentResultRepository from '../../../src/shared/infrastructure/repositories/assessment-result-repository.js';
-import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
-import * as skillRepository from '../../../src/shared/infrastructure/repositories/skill-repository.js';
-import * as knowledgeElementRepository from '../../infrastructure/repositories/knowledge-element-repository.js';
-import * as knowledgeElementSnapshotRepository from '../../infrastructure/repositories/knowledge-element-snapshot-repository.js';
-import { PlacementProfile } from '../models/PlacementProfile.js';
-import { UserCompetence } from '../models/UserCompetence.js';
+} from '../../../certification/shared/domain/models/CertificationVersion.js';
+import * as scoringService from '../../../evaluation/domain/services/scoring/scoring-service.js';
+import * as assessmentRepository from '../../infrastructure/repositories/assessment-repository.js';
+import * as assessmentResultRepository from '../../infrastructure/repositories/assessment-result-repository.js';
+import * as competenceRepository from '../../infrastructure/repositories/competence-repository.js';
+import * as skillRepository from '../../infrastructure/repositories/skill-repository.js';
 
 async function getPlacementProfile({
   userId,
